requirements.txt
asgiref==3.3.1
backcall==0.2.0
blessings==1.7
coverage==5.3
decorator==4.4.2
dj-database-url==0.5.0
Django==3.1.4
django-configurations==2.2
django-filter==2.4.0
django-model-utils==4.1.1
django-nose==1.4.7
django-unique-upload==0.2.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
factory-boy==3.1.0
Faker==5.0.2
flake8==3.8.4
gunicorn==20.0.4
ipdb==0.13.4
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
Markdown==3.3.3
mccabe==0.6.1
mock==4.0.3
nose==1.3.7
nose-progressive==1.5.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
ptyprocess==0.6.0
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.7.3
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
traitlets==5.0.5
wcwidth==0.2.5
